feat(navbar): style and add navbar
