+++ app/helpers/boxset_helper.rb
@@ -1,0 +1,2 @@
+module BoxsetHelper
+end
