+++ app/helpers/input_helper.rb
@@ -1,7 +1,7 @@
 module InputHelper
   def input_styles
-    "w-full px-4 py-2 border border-gray-300 dark:border-gray-800 rounded-xl \n
-    bg-white dark:bg-gray-800 text-cyan-600 dark:text-gray-200 focus:border-none \n
-    focus:ring-1 focus:ring-nineGreen-500"
+    "w-full px-4 py-2 border border-gray-800 rounded-xl \n
+    bg-background text-gray-200 focus:border-none \n
+    focus:ring-1 focus:ring-accent-50"
   end
 end
