+++ app/views/boxsets/index.html.erb
@@ -24,11 +24,6 @@
       </div>
     </div>
   </div>
-  <% if boxset.nil? %>
-    <div class='p-8 rounded-lg bg-dark-200'>
-      <h3 class='text-white'>Select a boxset to view cards</h3>
-    </div>
-  <% end %>
   <div data-filter-dropdown-target="tableContainer" id="table-container" class='min-w-full px-16 mt-8 bg-dark-50'>
     <!-- Table will be loaded here -->
   </div>
