+++ app/views/boxsets/load_boxset.turbo_stream.erb
