+++ app/views/magic_cards/index.html.erb
@@ -1,0 +1,1 @@
+hi
