+++ app/views/revise_auth/registrations/new.html.erb
@@ -1,5 +1,5 @@
 <div class='flex items-start justify-center w-full min-h-screen pt-[160px]'>
-  <div class='w-full max-w-md p-6 bg-white shadow-lg rounded-xl'>
+  <div class='w-full max-w-md p-6 shadow-lg bg-foreground rounded-xl'>
     <h1 class='mb-2 text-xl font-bold text-center text-nineGreen-500'>Create your account</h1>
     <h1 class='mb-6 text-sm text-center '>Start making your dreams come true</h1>
     <%= form_with model: @user, url: sign_up_path, class: "space-y-4" do |form| %>
@@ -47,7 +47,7 @@
         <% end %>
       </div>
       <div class='flex flex-col items-center w-full'>
-        <%= form.submit "Create Account", class: "w-full py-2 px-4 mt-2 bg-nineGreen-500 text-white rounded-md hover:bg-nineGreen-600 transition duration-300" %>
+        <%= form.submit "Create Account", class: "w-full py-2 px-4 mt-2 bg-accent-50 text-white cursor-pointer rounded-md hover:bg-nineGreen-600 transition duration-300" %>
         <%= link_to ("Log in"), login_path, class: 'mt-4' %>
       </div>
     <% end %>
