+++ app/views/revise_auth/sessions/new.html.erb
@@ -1,5 +1,5 @@
 <div class='flex items-start justify-center w-full min-h-screen pt-[160px]'>
-  <div class='w-full max-w-md p-6 bg-white shadow-lg rounded-xl'>
+  <div class='w-full max-w-md p-6 border border-foreground bg-foreground rounded-xl text-greyText'>
     <h1 class='mb-2 text-xl font-bold text-center text-nineGreen-500'>Log in to your account</h1>
     <h1 class='mb-6 text-sm text-center '>Start making your dreams come true</h1>
     <%= form_with url: login_path, class: "space-y-4" do |form| %>
@@ -17,7 +17,7 @@
             <%= flash[:alert] %>
           </div>
         <% end %>
-        <%= form.submit "Log in", class: "w-full py-2 px-4 mt-2 bg-nineGreen-500 text-white rounded-md hover:bg-nineGreen-600 transition duration-300" %>
+        <%= form.submit "Log in", class: "w-full py-2 px-4 mt-2 bg-accent-50 text-background rounded-md cursor-pointer hover:bg-nineGreen-600 transition duration-300" %>
         <%= link_to ("Sign up"), sign_up_path, class: 'mt-4' %>
       </div>
     <% end %>
