+++ app/views/shared/_navbar.html.erb
@@ -1,0 +1,23 @@
+<div class='flex flex-row h-[72px] bg-menu px-64'>
+  <div class='flex flex-row items-center justify-between w-full space-x-4 text-white'>
+    <div class='flex flex-row items-center space-x-2'>
+      <p class='pr-6'>99 Staples</p>
+      <div class="<%= nav_item_classes('/magic_cards') %>">
+        <%= link_to "Home", magic_cards_path %>
+      </div>
+      <div class="<%= nav_item_classes('/boxsets') %>">
+        <%= link_to "Boxsets", boxsets_path %>
+      </div>
+      <p class="<%= nav_item_classes('') %>">Collections</p>
+      <p class="<%= nav_item_classes('') %>">Decks</p>
+    </div>
+    <div class='flex flex-row space-x-6'>
+      <% if current_user %>
+        <%= button_to 'Start Ingest Process', dashboard_ingest_path, method: :post, class: '' %>
+        <%= link_to "Sign out", logout_path, data: { turbo_method: :delete }, class: "" %>
+      <% else %>
+        <%= link_to "Log In", login_path, class: "" %>
+      <% end %>
+    </div>
+  </div>
+</div>
